+++ example/lib/main.dart
@@ -24,8 +24,8 @@ Map<String, IconData> materialSymbolsOutlinedMap = {};
 Map<String, IconData> materialSymbolsRoundedMap = {};
 Map<String, IconData> materialSymbolsSharpMap = {};
 
-const String materialSymbolsIconsSourceFontVersionNumber = '2.777';  // must update for each new font update
-const String materialSymbolsIconsSourceReleaseDate = 'August 1, 2024';  // must update for each new font update
+const String materialSymbolsIconsSourceFontVersionNumber = '2.780';  // must update for each new font update
+const String materialSymbolsIconsSourceReleaseDate = 'August 9, 2024';  // must update for each new font update
 int totalMaterialSymbolsIcons=0;
 
 void makeSymbolsByStyleMaps() {